[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -9,7 +9,7 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    int A, B;
+    long long A, B;
     cin >> A >> B;
     cout << (A + B) * (A - B) << endl;
 }
